backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
llama-index==0.9.10
langchain==0.0.340
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.1
pgvector==0.2.3
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-dotenv==1.0.0
redis==5.0.1
numpy==1.24.3
tenacity==8.2.3